+++ src/integrationTest/resources/response/bulk-scan-fl401-transform-output.json
@@ -16,28 +16,28 @@
           },
           {
             "value": {
-              "fullName": "Bill",
-              "dateOfBirth": " 03/04/08",
+              "fullName": "Will",
+              "dateOfBirth": " 04/05/11",
               "applicantChildRelationship": " father",
-              "applicantRespondentShareParental": " yes",
+              "applicantRespondentShareParental": " no",
               "respondentChildRelationship": " mother"
             }
           },
           {
             "value": {
-              "fullName": "Bill",
-              "dateOfBirth": " 03/04/08",
-              "applicantChildRelationship": " father",
-              "applicantRespondentShareParental": " yes",
-              "respondentChildRelationship": " mother"
+              "fullName": "Ash",
+              "dateOfBirth": " 11/05/16",
+              "applicantChildRelationship": " Mother",
+              "applicantRespondentShareParental": " no",
+              "respondentChildRelationship": " father"
             }
           },
           {
             "value": {
-              "fullName": "Bill",
-              "dateOfBirth": " 03/04/08",
+              "fullName": "Jerry",
+              "dateOfBirth": " 12/05/17",
               "applicantChildRelationship": " father",
-              "applicantRespondentShareParental": " yes",
+              "applicantRespondentShareParental": " no",
               "respondentChildRelationship": " mother"
             }
           }
